프로젝트 연결해제시 RCS 등록된 정보에 한해서 RCS 삭제 API 호출

+++ src/main/java/kr/co/uplus/cm/project/controller/ProjectController.java
@@ -712,7 +712,7 @@ public class ProjectController {
 		RestResult<Object> rtn = new RestResult<Object>();
 
 		try {
-			if ("Y".equals(params.get("delYn").toString())) {
+			if ("Y".equals(params.get("delYn").toString()) && params.get("chatbotId") != null) {
 				projectService.deleteCallbackForApi(params);
 			} else {
 				projectService.delCallNum(params);
